requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-common-helpers==0.9.2
django-crontab==0.7.1
docker==4.4.1
idna==2.10
idna-ssl==1.1.0
maxminddb==2.0.3
maxminddb-geolite2==2018.703
multidict==5.1.0
path==15.0.1
pkg-resources==0.0.0
psutil==5.8.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.6.3
